pyproject.toml
[project]
name = "36-elemental-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.2",
    "questionary>=2.1.0",
    "rich>=14.0.0",
    "streamlit>=1.45.0",
]
